+++ massiv-bench/bench/Stencil.hs
@@ -1,105 +1,75 @@
 {-# LANGUAGE BangPatterns #-}
 {-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE GADTs #-}
 module Main where
 
-import Bench
-import Bench.Massiv as A
+import Data.Massiv.Array as A
+import Data.Massiv.Array.Stencil.Unsafe as A
+import Data.Massiv.Array.Unsafe as A
+import Data.Massiv.Bench as A
 import Criterion.Main
-import Prelude as P
 
+avg3x3 :: Fractional a => (Ix2 -> a) -> a
+avg3x3 get =
+  (get (0 :. 0) + get (0 :. 1) + get (0 :. 2) +
+   get (1 :. 0) + get (1 :. 1) + get (1 :. 2) +
+   get (2 :. 0) + get (2 :. 1) + get (2 :. 2) ) / 9
+{-# INLINE avg3x3 #-}
 
+avg3x3Stencil :: (Default a, Fractional a) => Stencil Ix2 a a
+avg3x3Stencil = makeStencil (Sz (3 :. 3)) 0 avg3x3
+{-# INLINE avg3x3Stencil #-}
 
+avg3x3StencilUnsafe :: Fractional a => Stencil Ix2 a a
+avg3x3StencilUnsafe = makeUnsafeStencil (Sz (3 :. 3)) 0 (const avg3x3)
+{-# INLINE avg3x3StencilUnsafe #-}
 
+sum3x3StencilConv :: Fractional a => Stencil Ix2 a a
+sum3x3StencilConv = makeConvolutionStencil (Sz2 3 3) 0 $ \ get ->
+  get (0 :. 0) 1 . get (0 :. 1) 1 . get (0 :. 2) 1 .
+  get (1 :. 0) 1 . get (1 :. 1) 1 . get (1 :. 2) 1 .
+  get (2 :. 0) 1 . get (2 :. 1) 1 . get (2 :. 2) 1
+{-# INLINE sum3x3StencilConv #-}
+
+sum3x3StencilConvKern :: (Prim a, Fractional a) => Stencil Ix2 a a
+sum3x3StencilConvKern =
+  makeConvolutionStencilFromKernel $ makeArrayR P Seq (Sz2 3 3) (const 1)
+{-# INLINE sum3x3StencilConvKern #-}
 
 main :: IO ()
 main = do
-  let t2 = (1600, 1200) :: (Int, Int)
+  let sz = Sz2 1600 1200
   defaultMain
     [ bgroup
         "Stencil"
-        [ bgroup
-            "Average Seq"
-            [ env
-                (return (computeAs U (arrDLightIx2 Seq (tupleToIx2 t2))))
-                (bench "Array Ix2" .
-                 whnf (computeAs U . A.mapStencil (Fill 0) average3x3Filter))
-            , env
-                (return (computeAs U (arrDLightIx2 Seq (tupleToIx2 t2))))
-                (bench "Unsafe Array Ix2" .
-                 whnf (computeAs U . average3x3FilterUnsafe))
-            , env
-                (return (computeAs U (arrDLightIx2 Seq (tupleToIx2 t2))))
-                (bench "Convolve Array Ix2" .
-                 whnf (computeAs U . A.mapStencil (Fill 0) average3x3FilterConv))
-            ]
-        , bgroup
-            "Average Par"
-            [ env
-                (return (computeAs U (arrDLightIx2 Par (tupleToIx2 t2))))
-                (bench "Array Ix2" .
-                 whnf (computeAs U . A.mapStencil (Fill 0) average3x3Filter))
-            , env
-                (return (computeAs U (arrDLightIx2 Par (tupleToIx2 t2))))
-                (bench "Convolve Array Ix2" .
-                 whnf (computeAs U . A.mapStencil (Fill 0) average3x3FilterConv))
-            ]
+        [ env (return (arrRLightIx2 P Seq sz)) $ \arr ->
+            bgroup
+              "Average Seq"
+              [ bench "Array Ix2" $ whnf (computeAs P . A.mapStencil (Fill 0) avg3x3Stencil) arr
+              , bench "Array Ix2" $
+                whnf
+                  (computeAs P . A.applyStencil (samePadding avg3x3Stencil (Fill 0)) avg3x3Stencil)
+                  arr
+              , bench "Unsafe Stencil Array Ix2" $
+                whnf (computeAs P . A.mapStencil (Fill 0) avg3x3StencilUnsafe) arr
+              , bench "Unsafe Map Array Ix2" $
+                whnf (computeAs P . A.mapStencilUnsafe (Fill 0) 3 0 avg3x3) arr
+              , bench "Convolve Array Ix2" $
+                whnf (computeAs P . A.mapStencil (Fill 0) (sum3x3StencilConv / 9)) arr
+              , bench "Convolve Kernel Array Ix2" $
+                whnf (computeAs P . A.mapStencil (Fill 0) (sum3x3StencilConvKern / 9)) arr
+              , bench "Foldl Array Ix2" $
+                whnf (computeAs P . A.mapStencil (Fill 0) (foldlStencil (+) 0 (Sz 3) / 9)) arr
+              , bench "Monoid Array Ix2" $
+                whnf (computeAs P . A.mapStencil (Fill 0) (avgStencil (Sz 3))) arr
+              ]
+        , env (return (arrRLightIx2 P Par sz)) $ \arr ->
+            bgroup
+              "Average Par"
+              [ bench "Array Ix2" $ whnf (computeAs P . A.mapStencil (Fill 0) avg3x3Stencil) arr
+              , bench "Convolve Array Ix2" $
+                whnf (computeAs P . A.mapStencil (Fill 0) (sum3x3StencilConv / 9)) arr
+              , bench "Monoid Array Ix2" $
+                whnf (computeAs P . A.mapStencil (Fill 0) (avgStencil (Sz 3))) arr
+              ]
         ]
-    -- , bgroup
-    --     "Sobel"
-    --     [ bgroup
-    --         "Horizontal"
-    --         [ env
-    --             (return (massDLightIx2 Seq (tupleToIx2 t2)))
-    --             (bench "Massiv Ix2 U" .
-    --              whnf (M.mapStencil (sobelX Edge)))
-    --         , env
-    --             (return (computeAs U (arrDLightIx2 Seq (tupleToIx2 t2))))
-    --             (bench "Array Ix2 U" .
-    --              whnf (computeAs U . A.mapStencil (sobelX Edge)))
-    --         , env
-    --             (return (computeUnboxedS (arrDLightSh2 (tupleToSh2 t2))))
-    --             (bench "Repa DIM2 U" . whnf (computeUnboxedS . mapSobelRX))
-    --         ]
-    --     , bgroup
-    --         "Vertical"
-    --         [ env
-    --             (return (computeAs U (arrDLightIx2 Seq (tupleToIx2 t2))))
-    --             (bench "Array Ix2 U" .
-    --              whnf (computeAs U . A.mapStencil (sobelY Edge)))
-    --         , env
-    --             (return (computeUnboxedS (arrDLightSh2 (tupleToSh2 t2))))
-    --             (bench "Repa DIM2 U" . whnf (computeUnboxedS . mapSobelRY))
-    --         ]
-    --     , bgroup
-    --         "Operator Fused Seq"
-    --         [ env
-    --             (return (computeAs U (arrDLightIx2 Seq (tupleToIx2 t2))))
-    --             (bench "Array Ix2 U" .
-    --              whnf (computeAs U . A.mapStencil (sobelOperator Edge)))
-    --         , env
-    --             (return (computeUnboxedS (arrDLightSh2 (tupleToSh2 t2))))
-    --             (bench "Repa DIM2 U" . whnf (computeUnboxedS . sobelOperatorR))
-    --         ]
-    --     , bgroup
-    --         "Operator Fused Par"
-    --         [ env
-    --             (return (computeAs U (arrDLightIx2 Par (tupleToIx2 t2))))
-    --             (bench "Array Ix2 U" .
-    --              whnf (computeAs U . A.mapStencil (sobelOperator Edge)))
-    --         , env
-    --             (return (computeUnboxedS (arrDLightSh2 (tupleToSh2 t2))))
-    --             (bench "Repa DIM2 U" .
-    --              whnf (runIdentity . computeUnboxedP . sobelOperatorR))
-    --         ]
-    --     , bgroup
-    --         "Operator Unfused"
-    --         [ env
-    --             (return (computeAs U (arrDLightIx2 Seq (tupleToIx2 t2))))
-    --             (bench "Array Ix2 U" . whnf (sobelOperatorUnfused Edge))
-    --         , env
-    --             (return (computeUnboxedS (arrDLightSh2 (tupleToSh2 t2))))
-    --             (bench "Repa DIM2 U" . whnf sobelOperatorRUnfused)
-    --         ]
-    --     ]
     ]